AI_Services/app/core/config.py
import boto3
import os
from dotenv import load_dotenv

load_dotenv()

S3_BUCKET=os.getenv("AWS_BUCKET_NAME")
S3_REGION = os.getenv("AWS_REGION")

s3_client = boto3.client(
    "s3",
    aws_access_key_id=os.getenv("AWS_ACCESS_KEY"),
    aws_secret_access_key=os.getenv("AWS_SECRET_ACCESS_KEY"),
    region_name=S3_REGION
)