AI_Services/requirements.txt
langchain-google-genai
python-dotenv
langchain
matplotlib
pydantic
langchain_huggingface
langchain_core
langchain_community
sentence-transformers
pinecone
langchain_pinecone
llama-cloud>=1.0
nest_asyncio
pymupdf
PDFPlumber
fastapi 
uvicorn 
streamlit 
requests
python-multipart
sqlalchemy
asyncpg
psycopg[binary]
boto3
langchain-redis